mxm: fix anysource_matched
The return value of anysource_matched should be the actual result
of the cancel operation. If the result is uncancelable, i.e. already
matched, then CH3 will let the netmod message win and move on to the
other requests in the queue. When the completion for the unsuccessfully
canceled message comes in, we process it like normal.

Reviewed-by: Igor Ivanov <Igor.Ivanov@itseez.com>
Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>