Added support for static compilation.
If the user passes the -static flag, we disable interlibrary
dependencies, since the linker would pick the static versions of the
libraries in that case.

We are still making an assumption that the default mode of the linker
is shared, and the user can specify static library builds with
-static.  However, this is not always true.  On BG/Q, for example, the
default is static.

Fixes #2190.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>