Bug-fix: add true_lb when directly issuing contiguous derived data.
In RMA communication, when data to be sent is derived datatype but
contiguous, we can avoid copy/packing but directly issue it from
the user buffer. However, the data may have non-zero lb and we should
add the true lb when calculating the starting address. This patch
fixes this issue.

Signed-off-by: Pavan Balaji <balaji@anl.gov>