Fix handle leaks when canceling LMT send requests
If the destination of a LMT nonblocking send, which
follows a rendez-vous protocol, receives the RTS packet to the
unexpected queue followed by a cancelation request, the request in
the unexpected queue never gets freed. To solve the issue, we
forcefully free the rendez-vous request by adding an additional
request release operation. Refs #287

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>