Correct usage of req's segment_first and segment_size in Nemesis
Originally the implementation of sendNoncontig() in Nemesis
assume that req->dev.segment_first is 0 and req->dev.segment_size
is the size of data, which potentially asssumes that the data passed
from upper layer must be the entire data, prohibiting the possibility
of streaming that data in the upper layer. The general way to use them
should be that req->dev.segment_first specifies the current starting
location and req->dev.segment_size specifies the current ending location.

We fixed this issue in commit 5132e070e0499b3a6fa2955710bef4c699d531fc, but
there are some places missed in that fixing. Here we fixed those places.

Signed-off-by: Pavan Balaji <balaji@anl.gov>