OFI netmod: RCD protocol bug fix.
The previous version of the protocol assumed that fi_trecv utilizes
direct address for matching. This does not hold, for example, for PSM
OFI provider (it does not set claim FI_DIRECTED_RECV capability). So,
the API1 set should work w/o directed recv assumption. Hence, we have
to encapsulate rank info in the tag in the RCD protocol in order to
avoid mismatch.

Change-Id: I7868f2f033cf3cdec29f7daded8b29c275b2632f
Signed-off-by: Charles J Archer <charles.j.archer@intel.com>