Check if type_tag is agnostic to modifiers.
Add a test to configure.ac that tests, if type_tag_for_datatype is
agnostic for modifiers as const or volatile.  If the test fails, the
type_tag_for_datatype test is disabled in mpi.h by setting a flag to
avoid compiler warning.  In this case, there are no type tests for MPI
data types when compiling.

Fixes #2223.

Signed-off-by: Pavan Balaji <balaji@anl.gov>