Handling of discard connection to avoid reconnect
The loser of a head-to-head connection sometimes tries
to reconnect later, afer MPI_Finalize was called  This
can lead to several errors in the socket layer, depending
on the state of the disarded connection and the appereance
of the connection events. Refs #2180
This Patch has two ways to handle this:

1.)
Discarded connections are marked with CONN_STATE_DISCARD,
so they are hold from connection.  Furthermore, an error on
any discarded connection (because the remote side closed in
MPI_Finalize) is ignored and the connection is closed.

2.)
Add a finalize flag for process groups. If a process group is
closing and tries to close all VCs, a flag is set to mark this.
If the flag is set, a reconnection (in the socket state) is
refused and the connection is closed on both sided.

Both steps are necessary to catch all reconnection tries after
MPI_Finalize was called.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>