Bug-fix: free lock entry queue when it was allocated.
Originally we allocate the lock entry queue in Win_create/allocate
(if no_locks is not set), and free the queue in Win_free (if no_locks
is not set). This is not correct because no_locks may be set after
Win_create/allocate and before Win_free. In this patch, we free
the queue in Win_free if the queue was allocated in Win_create/allocate.

Thanks to Lisandro Dalcin [dalcinl@gmail.com] for reporting this bug.

Fix #2273

Signed-off-by: Pavan Balaji <balaji@anl.gov>