Move outstanding_acks increment to flush sending step.
The outstanding_acks counter was increased at each sync call (such as
fence and flush). However, the counter had to be decreased again if
flush ack is not required. It is more straightforward if increasing it
only when the flush packet is issued (FLUSH flag piggyback or a separate
flush message).

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>
Signed-off-by: Pavan Balaji <balaji@anl.gov>