Test send cancellation issues with unmatched sends
In this test, two messages are sent with the same rank, comm, and tag,
and one of them is cancelled before it gets matched. If the
destination does not differentiate between the messages, e.g. by using
sequence numbers or the origin request handle, the wrong message might
be canceled.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>