Perf-tuning: avoid dynamic allocation of requests in RMA.
Originally we always dynamically allocate a request array
for the current RMA operation, since the current operation
might be streamed and needs multiple requests to track
each stream unit. However, in most cases where streaming is
not happening, we only needs one request for each operation
and does not need to dynamically allocate it. This patch
optimizes such case.

Signed-off-by: Pavan Balaji <balaji@anl.gov>