Delete poke_progress_engine in RMA sync calls and only keep self lock cases.
Originally we poke the progress engine at the end of RMA sync calls
if progress engine is never poked in this call before. The purpose
of this is to prevent possible deadlock problem. However, the deadlock
problem should only happen in self lock cases, if target is not myself,
it add unnecessary overhead to RMA sync calls. In this patch, we delete
those progress poking but only leave ones when target is myself.

Signed-off-by: Pavan Balaji <balaji@anl.gov>