Adding reduce-scatter based algorithm in MPI_Win_fence.
In this patch, we add a reduce-scatter based algorithm in
MPI_Win_fence, which is triggered when number of processes
is at a small / medium value. When this algorithm is being
used, memory usage is O(P), but the ending FENCE only needs
to wait for local completion but does not need to wait for
remote completion. When number of processes is large, we
switch FENCE to the original barrier based algorithm, which
has O(1) memory usage, but needs to wait for the remote
completion in the ending FENCE.

Signed-off-by: Pavan Balaji <balaji@anl.gov>