better approach for do_accumulate_op
commit 83253a414 triggerd a bunch of new warnings.  Take a different
approach.  For simplicity of implementation, do_accumulate_op is defined
as MPI_User_function.  We could split up internal routine and
user-provided routines, but that complicates the code for little
benefit:

Instead, keep do_accumlate_op with an int type, but check for overflow
before explicitly casting.  In many places the count is simply '1'.  In
stream processing there is an interal limit of 256k, so the assertion
should never fire.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>