Bug-fix: correct SHM memory size in Nemesis Win_gather_info().
In the Nemesis implementation of Win_gather_info(), we allocate
a memory region on SHM to store window information for other
processes, so that all processes on the same node can share
those information. However, previously the memory size was
incorrectly set as O(node_comm_size), which should be O(comm_size).
This patch fixed this bug.

Signed-off-by: Min Si <msi@il.is.s.u-tokyo.ac.jp>
Signed-off-by: Pavan Balaji <balaji@anl.gov>