Bug-fix: correct local/remote completion in Win_flush_local.
The original implementation in Win_flush_local counts number
of total local completion and remote completion needed to
wait, and then waiting for current local/remote completion
count to reach those values. There is a bug that we should
initialize the current count to zero in each while loop,
otherwise the targets that are already completed will be
count again and we failed to wait for some targets to be
completed. This patch fixes this issue.

Signed-off-by: Pavan Balaji <balaji@anl.gov>