Bug-fix: set kind of user request in request-based RMA operation correctly.
Here we should set the kind of user request in request-based RMA
operation to MPID_WIN_REQUEST, so that in MPI_Wait function, after
request is completed, it will decrement the ref counter of the
request and free it.

Originally the kind is set to MPID_REQUEST_SEND, which goes into
the path of MPID_REQUEST_SEND in MPI_Wait and causes segment fault
when --enable-debuginfo is set in configure.

Signed-off-by: Pavan Balaji <balaji@anl.gov>