Remove unnecessary asm checks for nemesis.
We were testing to see if we were on an x86 32-bit platform that
accepts some assembly code to optimize 32-bit architectures.  This is
only for 32-bit systems and seems to be not comprehensive enough to
avoid failures with some compilers (e.g., icc-13.1).  Since such
performance optimization on 32-bit platforms is not really a priority
for us, it's cleaner to just remove this.

Signed-off-by: Halim Amer <aamer@anl.gov>