Revert "Perf-tuning: overlapping issuing data and computation on target side in GACC/FOP."
This reverts commit 06dbf44b8fe8cc313b5bab53e2cfd59b8085f718.

In commit 06dbf44b8fe8cc313b5bab53e2cfd59b8085f718, for GACC/FOP/CAS,
the target side first starts sending back data, then performs computation,
in order to overlap data transmission and computation. This is not
correct when this operation needs to be piggybacked with a FLUSH flag,
or a Win_flush call is called after this operation, because the
completion of computation on the target side is not guaranteed.
Here we revert this patch so that data is sent back after computation
is done on the target side.