Replace @PERL@ in maint/ with "/usr/bin/env perl".
Scripts in maint/ are generated by autogen.sh which may not be executed
by end users. Thus @PERL@ may not be parsed to the correct path of perl.
This patch fixed this issue.

Note that we also replace "-w" with "use warning;" because env may
incorrectly parse "perl -w" as the application name.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>