Bugfix for MPI_Ibarrier for inter_communicators
MPI_Ibarrier for inter-communicators fails because
of a wrong condition check. In MPIR_Ibarrier_impl,
checking “comm_ptr->local_size != 1” is used for
speeding up MPI_Ibarrier for intra-communicators.
However, for inter-communicators, this prevents the
barrier to synchronize with the remote group and
cause hang. An inter-communicator still have to
synchronize with the remote group, even is the local
size is one.

In the patch, checking “local_size != 1” is now moved to
the MPIR_Ibarrier_inter when doing local sync.

Refs #2284

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>