Use request_completed_cb to complete user request in req-based RMA operation.
Originally we use OnDataAvail function pointer in the request
to complete the user request when the operation is completed,
which is not a good choice since the function pointer should
be used in send/recv when the next data chunk is available.
Here we use request_completed_cb function pointer to complete
the user request.

Signed-off-by: Pavan Balaji <balaji@anl.gov>