Adding a request callback function for ibarrier in Win_fence.
In Win_fence implementation, when we enter the second Win_fence
and request of the ibarrier in the first Win_fence has not been
completed yet, we do not need to wait for that request's
completion and do not need to store that request on window
anymore. However, we do need to keep track of number of such
incomplete but 'out-of-date' requests on the window in order
to clean them up in Win_free.

In this patch, we achieve this by adding a counter for incomplete
but 'out-of-date' requests and a callback function to those requests.
We increment the counter when we meet one such request, and
decrement the counter in the callback function when that request
is got completed. By doing so, Win_free can just block waiting for
that counter to become 0 to clean up all pending requests.

Signed-off-by: Pavan Balaji <balaji@anl.gov>