Adds tests for threaded collective operations
A new test directory test/mpi/threads/coll is created and two new tests
are added to this directory. The new tests will test threaded overlapped
collective operations.

A comprehensive test of these two new test case is still needed. For now,
the new tests are known to fail with ch3:socket so they are marked as xfail.
Refs #2108

Signed-off-by: Lena Oden <loden@anl.gov>