Move commit_comm to the end of the comm_idup
The function commit_comm was called for comm_idup
before the new context_id was allocated. This leads to the
problem that the network-hook (mxm) is called with a
wrong/fake context_id. Now the function is scheduled at the
end, when all processes have found the correct context_id
and recvcontext_id.

Fixes #2283

Signed-off-by: Pavan Balaji <balaji@anl.gov>