Use sync_request_cnt to count total number of ibarrier requests.
Originally we use this counter to track number of dangling requests in
Fence.  Here we modify it to track total number of Fence requests,
which is a cleaner design.  This has the disadvantage that we now wait
for all previously pending ibarrier requests to complete before
issuing the pending operations, but that's not a common use case and
this patch is simpler and easier to maintain.

Signed-off-by: Pavan Balaji <balaji@anl.gov>