Add a callback for PSCW request completions.
Here the callback function will activate/de-activate the
window after all requests for PSCW are completed.

We also delete the activation/de-activation code in the opening
and the ending synchronization calls for PSCW which were
originally added for debugging.

When a new op is enqueued / all operations are issued out, we
activate the window / de-activate the window under PSCW state.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>