Set window to active when receiving Lock ACK.
Here we activate th window when lock is granted and
there are pending operations.

We also delete the activation/de-activation code in the opening
and the ending synchronization calls for Passive Target which were
originally added for debugging.

When a new op is enqueued / all operations are issued out, we
activate the window / de-activate the window under Passive
Target states.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>