Remove unused old code.
When we added the new alltoall code, we commented out the old code,
but kept it in case there were bugs in the new code.  It's been
several years since we did that and haven't noticed any bugs.  So it's
safe to delete the old code.  It's safely stored in the version
control system anyway.

Signed-off-by: Lena Oden <loden@anl.gov>