Modify Fence cb function to activate/de-activate window.
Same with PSCW, after all previous ibarrier requests are
completed, the cb function will activate/de-activate the
current window, depending on if there are pending operations.

We also delete the activation/de-activation code in the opening
and the ending synchronization calls for FENCE which were
originally added for debugging.

When a new op is enqueued / all operations are issued out, we
activate the window / de-activate the window under FENCE state.

Signed-off-by: Pavan Balaji <balaji@anl.gov>