Laying foundation for new features for Panasas
New versions of Panasas (Panfs-6 and up) implement RAID0 and RAID5 in
terms of newer RAIDN support, but Panfs-5 and older don't know what
RAIDN is.  This new 'open6' will in subsequent patches make use of new
features, but we'll keep the old open around in case sites have not
upgraded.