Split pending op list into pending_net_ops_list and pending_user_ops_list.
Here we split pending op list in the target object into two lists:
(1) pending_net_ops_list and (2) pending_user_ops_list.

(1) is used to store operations that are held to wait for network events
(synchronization), and (2) is used to store operations that are held to
wait for user events (MPI_Win_unlock/flush/flush_local posted). When
user events happen, we move the operation in (2) to (1).

Signed-off-by: Pavan Balaji <balaji@anl.gov>