Bug-fix when Lock Ack is LOCK_QUEUED_DATA_QUEUED.
When flag in Lock Ack is LOCK_QUEUED_DATA_QUEUED, it means that
the entire operation is queued on the target side, at this point,
we should not complete and delete that operation. When origin side
receives Lock Ack with LOCK_GRANTED, it will delete that operation
from the list.

Signed-off-by: Pavan Balaji <balaji@anl.gov>