move static function check_and_set_req_completion to header file.
This static function checks the completion of requests in RMA
operation and sets completion cb function if requests are not
completed. Originally it was only used in RMA issuing progress,
now we need to use it in handle_lock_with_op as well, so here
we move it to the header file.

Signed-off-by: Pavan Balaji <balaji@anl.gov>