Free RMA op object immediately after it is issued out.
Originally we maintain RMA operations after they are
issued out in order to detect the local completion of
those operations, since operation structure will be
freed in the request completion cb when operation is
locally completed. However, we do not need to maintain
operation structure at all after the operation is
issued out, instead, we can just use a counter to keep
track of number of operations that are waiting for
local completion, and decrement that counter in the
request completion cb.

In this patch, we delete all the code that related
to maintain ssued but incomplete operation structure.
Therefore, operation objects exist only when synchronization
is not finished and operations stay in the pending list.

Signed-off-by: Pavan Balaji <balaji@anl.gov>