Use counter num_ops_flush_not_issued to control issuing of FLUSH messages.
Here we use counter 'num_ops_flush_not_issued' to count number of operations
that are issued out but no FLUSH message issued afterwards yet to detect
their remote completion. When we reach the tail of the pending list, if
this counter is non-zero, we will issue a FLUSH message and reset the
counter to zero; if the counter is zero, which means there is no operation
that needs detecting remote completion, then we do not need to issue FLUSH
message at all.

Originally we use a flag 'put_acc_issued' to achieve the similiar purpose,
but that is not a clean design and miss some cases. This counter is a better
design on saving FLUSH messages as many as possible.

Signed-off-by: Pavan Balaji <balaji@anl.gov>