Reduce default value of MPIR_CVAR_CH3_RMA_ACTIVE_REQ_THRESHOLD.
The original default value of MPIR_CVAR_CH3_RMA_ACTIVE_REQ_THRESHOLD
is too high and can cause the internal resources to be used up when
all processes within one node have lots of internal requests. Here
we reduce this default value.

We set the value by meansuing one-to-all communication between MXM
nodes, when one origin process issues operations to 8 targets. The
performance achieves steady state when the origin processes issues
~16K operations to each target, which means the network is saturated
at 16K operations. Here, instead of setting value as 16K*8, we set
the default value as 64K, trying to keep the resources used by
requests as small as possible. Note that in future after we add the
support for hardware RDMA operations to RMA infrastructure, the
pressure on request resources should be reduced a lot.

Signed-off-by: Pavan Balaji <balaji@anl.gov>