Increase time limit of some GACC-based RMA tests to 4 min.
In 40b1483fb31d9207644783a2574b08149dfa594b, we reduced the value
of MPIR_CVAR_CH3_RMA_ACTIVE_REQ_THRESHOLD to 64K by measuring with
an one-to-all benchmark involving PUT operations on MXM. This effects
the performance of some GACC-based tests. Here we increase the time
limit of those tests to 4 min.

The default value of MPIR_CVAR_CH3_RMA_ACTIVE_REQ_THRESHOLD is just
a reference value according the one-to-all PUT tests. In a real
application, user should tune the value depending on different
operations and communication patterns.

Signed-off-by: Pavan Balaji <balaji@anl.gov>