Bug-fix: Handle icc failures on Darwin with thread local storage.
The intel compiler seems to break on Darwin when the thread local
storage is specified with the extern keyword.  With older intel
compilers, which used to break more easily for tls specifiers, we
added a configure check to disable tls in such cases.  The newer
compilers seem to work fine with no extern is provided, but break when
extern is provided for shared library builds.  This patch improves the
configure check to try both with and without extern.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>