First round of cleanup for ROMIO threading.
ROMIO uses the same macro names as MPICH, but with different
semantics.  This causes a lot of confusion as to what it really does.
This patch simply cleans up the usage and gives the macro a different
name.  It also removes the unnecessary parameters since they are
simply ignored anyway.  Having no parameters makes it clearer to
future developers that these macros do not require any arguments.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>