Orders multiple comm_idups in a list
This patch modifies the handling of multiple MPI_Comm_idup.
Instead of using a sequence number for every communicator
to handle the ordering of comm_idups, now a linked list is used.
Only the first comm_idup in the list can own the mask.

The comm_idup operations are ordered in the list using the
context id of the parent communicator.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>