link MPL with included process managers
Process managers gforker, remshell, and their generic utils all rely on
MPL functionality, so we add explicit links in each Makefile.mk. In order
for libtool to handle the heavy lifting when linking MPL with libmpiexec,
we convert libmpiexec to a libtool convenience library.

To support both embedded and external MPL libraries, we always define
mpllibdir in configure.ac and add to LDFLAGS where needed.

Fixes #2299

Signed-off-by: Rob Latham <robl@mcs.anl.gov>