ch3/vc: changes to node id mapping
Failures occurred with several of the spawn tests when run across
multiple nodes. The problem was the result of the node id mapping for
the spawned processes. If the spawned processes launched on only a
subset of nodes, they incorrectly reported a maximum node id in
MPID_Get_max_node_id.

Our approach to fix this is to report the maximum node id as simply
the largest id found during mapping. All references to the value are
updated to reflect it as the maximum node id, not a calculation of
the number of nodes, as that was inaccurate.

Reported-by: Artem Yalozo <artem.v.yalozo@intel.com>
Signed-off-by: Pavan Balaji <balaji@anl.gov>