maint/jenkins: add multi-node test scripts
Two scripts are added for multi-node tests. They are designed to be
submitted as two consecutive jobs.

mn-build.sh builds the MPI library and test binaries on local disk. The
temporary workspace on local disk is packed as a tarball and saved in
the Jenkins workspace. This script should be submitted as a 1-node SLURM
job using `srun`.

mn-run.sh recreates the temporary workspace and extracts the files from
the tarball in the previous step. Then it runs all the tests and
collects the results. This script should be submitted as a 2-node SLURM
job using `srun`.

Notes:
1. These two scripts must use the same temporary workspace location. The
location of temporary workspace is specified using "-t" option.
2. Each of these two scripts should only be ran one time. When submitted
through `srun`, SLURM will try to run these scripts on every allocated
nodes, which is incorrect. These scripts checks the SLURM_NODEID
variable and exit if it is not running on the first allocated node.
3. In order to specify interface for tcp jobs. We hack the
test/mpi/runtests script to include "-iface ib0" with mpiexec.

No reviewer.