New design for the ordering of context_id allocation for communicators
This patch changes the ordering-algorithm of the context_id allocation for communicators. Every
context_id allocation is added to a linked list after the first iteration. Only the first element in
this list can own the mask. This first iteration is required to avoid deadlocks. It ensures that
all processes have entered the function before the allocation-operation can be on top of the list.

The allocations are added to the list in order, according the context_id of the parent communicator
and a tag. This tag is user-defined for comm_group_create. For the non-blocking operations
(comm_idup) the scheduling tag is used to determine this tag. This is required to guarantee the
right ordering between different comm_idup operations on the same communicator - even if the
iallreduce operations of the first iteration are not completed in order.

Signed-off-by: Pavan Balaji <balaji@anl.gov>