binding/mpif_h: promote attribute copy argument types
MPI-2.1 added new attribute functions that changed the type of the
value and extra state arguments to be of MPI_ADDRESS_KIND in Fortran.
The mpif_h bindings, however, would truncate these values by casting
them as default Fortran integers when calling the underlying copy
functions. We address the issue by substituting MPI_Aint in place
of MPI_Fint for temporary variables in the binding code.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>