Wait for outstanding ACKs in ending synchronization calls.
When ordering of AM-based operations is not provided by network,
the runtime needs to add an ACK_REQ when issuing each operation,
so that any future internal FLUSH is guaranteed to be correct
(for example, when the runtime tries to reclaim resource). In the
ending synchronization, the runtime should wait for all ACKs to
come back before exit, no matter whether that synchronization is
waiting for local completion or remote completion.

Signed-off-by: Pavan Balaji <balaji@anl.gov>