maint/jenkins: xfail test failing on Portals 4 bug
The Portals 4 reference implementation has a bug that makes recv cancel
operations unreliable in threaded cases. We modified to the threaded
alltoall test to consistently reproduce the failure. We have xfailed it
when running with Portals 4 to not dirty the Jenkins results. Now we
apply the same xfail for any job that uses portals4 netmod.

No reviewer.