Add the use_rmk parameter to HYDT_bsci_launch_procs.
RMK is used to allocate nodes on a system before launching a job.  If
it is not specified in the user arguments for the UI process (e.g.,
mpiexec), it is set to the same as a job launcher (e.g., SLURM or
PBS).  This works fine for launching processes on nodes for the first
time, but it has a problem when a process dynamically creates other
processes at run time, e.g., by calling MPI_Comm_spawn(), because the
job launcher does not know which node(s) will be used for new
processes and it just allocates nodes based on its allocation policy.
This may conflict with the process management policy of the Hydra
framework, since the Hydra independently decides target nodes to
create a proxy process and to spawn new processes.  When the conflict
happens the spawned processes cannot communicate correctly.
To resolve this problem, a parameter 'use_rmk' is added to the
HYDT_bsci_launch_procs function.  If it is HYD_TRUE, RMK is used to
allocate nodes.  HYD_TRUE is passed when the HYDT_bsci_launch_procs
function is called from the UI process.  On the other hand, if it is
HYD_FALSE, we force not to use RMK.  HYD_FALSE is passed in PMI spawn
functions.

Signed-off-by: Pavan Balaji <balaji@anl.gov>