netmod/portals4: better check for datatype mismatch
The previous datatype mismatch check could generate false positives
when the received data was less than the total size of the posted recv
op. We take a new approach to detecting mismatches in the case where
datatypes are constructed from a single basic type. If the received
number of bytes is not a multiple of the size of a single basic element,
a mismatch has occurred.

Signed-off-by: Halim Amer <aamer@anl.gov>