remove use of send-side CQ tag field in ofi netmod
description:
The libfabric provider does not need to supply a valid tag field for
send-side events on a fi_cq_read() call.  The ofi netmod was relying
on this field, and all MPI tests will hang during finalize if it isn't
set.  The TCP libfabric provider had originally supplied a valid tag for
send events, but this has recently changed.  The gni provider does not
supply a valid tag field for send events.

author: Kim McMahon, Cray, Inc.

Signed-off-by: Charles J Archer <charles.j.archer@intel.com>