Make string CVAR case-insensitive
This patch fix ticket #2206. Currently, MPIR_CVAR_NEMESIS_NETMOD is
case-insensitive while MPIR_CVAR_DEFAULT_THREAD_LEVEL is not.  Replace
strcmp with strcasecmp for MPIR_CVAR_DEFAULT_THREAD_LEVEL to make the
comparison case-insensitive. Other CVAR of string type does not need to
be case-insensitive as they are for path and interface names.

Updating the description of case-insensitive CVARs.

Signed-off-by: Pavan Balaji <balaji@anl.gov>